[INTERNAL] static cast to double to avoid windows warning.

+++ include/seqan/arg_parse/arg_parse_version_check.h
@@ -288,7 +288,7 @@ inline bool _checkWritability(std::string const & path)
 
 inline double _getFileTimeDiff(std::string const & timestamp_filename)
 {
-    double curr = std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+    double curr = static_cast<double>(std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count());
     std::ifstream timestamp_file;
     timestamp_file.open(timestamp_filename.c_str());
